Call CloseBrowser to destroy browser window on Windows. Use a flag to dispatch close message to one window at a time. Add a menu bar for popup windows with File>Close instead of File>Exit used in the primary window.

+++ appshell/appshell_extensions_win.cpp
@@ -22,6 +22,7 @@
  */ 
 
 #include "appshell_extensions.h"
+#include "client_handler.h"
 
 #include <algorithm>
 #include <CommDlg.h>
@@ -29,6 +30,8 @@
 #include <Shlwapi.h>
 #include <stdio.h>
 #include <sys/stat.h>
+
+extern CefRefPtr<ClientHandler> g_handler;
 
 // Forward declarations for functions at the bottom of this file
 void ConvertToNativePath(ExtensionString& filename);
@@ -317,6 +320,11 @@ int32 SetPosixPermissions(ExtensionString filename, int32 mode)
 
 int32 DeleteFileOrDirectory(ExtensionString filename)
 {
+    DWORD dwAttr = GetFileAttributes(filename.c_str());
+
+    if ((dwAttr & FILE_ATTRIBUTE_DIRECTORY) != 0)
+        return ERR_NOT_FILE;
+
     if (!DeleteFile(filename.c_str()))
         return ConvertWinErrorCode(GetLastError());
 
@@ -325,7 +333,17 @@ int32 DeleteFileOrDirectory(ExtensionString filename)
 
 void CloseWindow(CefRefPtr<CefBrowser> browser)
 {
-    DestroyWindow(browser->GetHost()->GetWindowHandle());
+    if (browser.get() && g_handler.get()) {
+		g_handler->ClosingBrowser(true);
+	    browser->GetHost()->CloseBrowser();
+	}
+}
+
+void BringBrowserWindowToFront(CefRefPtr<CefBrowser> browser)
+{
+    HWND hwnd = browser->GetHost()->GetWindowHandle();
+    if (hwnd) 
+        ::BringWindowToTop(hwnd);
 }
 
 void ConvertToNativePath(ExtensionString& filename)
